Feat: auth_required로, doc과 jwt_required합친것 auth에 적용
