+++ app.py
@@ -42,7 +42,6 @@ def create_app():
     docs.register(mypage.get_look_items_info, blueprint=mypage.mypage.name)
     docs.register(auth.get_access_token, blueprint=auth.auth.name)
     docs.register(auth.refresh, blueprint=auth.auth.name)
-    docs.register(auth.hello, blueprint=auth.auth.name)
 
     docs.register(tinder.get_looks, blueprint=tinder.tinder.name)
     docs.register(tinder.confirm_look, blueprint=tinder.tinder.name)
